+++ app/src/main/AndroidManifest.xml
@@ -139,7 +139,7 @@
         </service>
 
         <activity
-            android:name=".ui.PagerActivity"
+            android:name=".ui.guide.PagerActivity"
             android:label="@string/app_name"
             android:theme="@style/AppTheme.NoActionBar">
             <intent-filter>
@@ -147,7 +147,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".MainActivity"
+            android:name=".ui.MainActivity"
             android:label="@string/app_name"
             android:theme="@style/AppTheme">
             <intent-filter>
