+++ app/src/main/java/com/github/simonpham/devtiles/ui/MainActivity.kt
@@ -1,7 +1,8 @@
-package com.github.simonpham.devtiles
+package com.github.simonpham.devtiles.ui
 
 import android.support.v7.app.AppCompatActivity
 import android.os.Bundle
+import com.github.simonpham.devtiles.R
 
 class MainActivity : AppCompatActivity() {
