+++ app/src/main/java/com/github/simonpham/devtiles/ui/guide/board/AddTileGuideFragment.kt
@@ -1,4 +1,4 @@
-package com.github.simonpham.devtiles.ui.guide
+package com.github.simonpham.devtiles.ui.guide.board
 
 import android.os.Bundle
 import android.support.v4.app.Fragment
@@ -6,18 +6,18 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import com.github.simonpham.devtiles.R
-import kotlinx.android.synthetic.main.fragment_welcome.*
+import kotlinx.android.synthetic.main.fragment_add_tile_guide.*
 
 /**
  * Created by Simon Pham on 6/1/18.
  * Email: simonpham.dn@gmail.com
  */
 
-class WelcomeFragment : Fragment() {
+class AddTileGuideFragment : Fragment() {
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?,
                               savedInstanceState: Bundle?): View? {
-        return inflater.inflate(R.layout.fragment_welcome, container, false)
+        return inflater.inflate(R.layout.fragment_add_tile_guide, container, false)
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
@@ -25,6 +25,6 @@ class WelcomeFragment : Fragment() {
 
         val listener = (activity as? View.OnClickListener)
                 ?: (parentFragment as? View.OnClickListener)
-        btnGettingStarted.setOnClickListener(listener)
+        btnDoneAddTile.setOnClickListener(listener)
     }
 }
