+++ app/src/main/java/com/github/simonpham/devtiles/ui/guide/board/StartDevelopingFragment.kt
@@ -1,4 +1,4 @@
-package com.github.simonpham.devtiles.ui.guide
+package com.github.simonpham.devtiles.ui.guide.board
 
 import android.os.Bundle
 import android.support.v4.app.Fragment
